changed way num_tickets called in lotteryq_choose, changed some types and names

+++ asgn2/runq.h
@@ -71,7 +71,7 @@ struct	thread *runq_choose(struct runq *);
 struct	thread *runq_choose_from(struct runq *, u_char);
 struct	thread *runq_choose_fuzz(struct runq *, int);
 struct  thread *lotteryq_choose(struct runq *, int);
-int 	rng(int);
+uint64_t rng(uint64_t);
 void	runq_init(struct runq *);
 void	runq_remove(struct runq *, struct thread *);
 void	runq_remove_idx(struct runq *, struct thread *, u_char *);
